Move shareLocation into LocationButton
